pyproject.toml
[tool.poetry]
authors = ["web3sdks <sdk@web3sdks.com>"]
description = ""
name = "web3sdks-sdk"
packages = [
  {include = "web3sdks"},
]
readme = "README.md"
version = "3.0.0"

[tool.poetry.dependencies]
python = ">=3.7.1"
dacite = "^1.6.0"
mypy-extensions = "^0.4.3"
web3sdks-contract-wrappers = "^2.0.4"
web3 = "5.27.0"
web3sdks-eth-account = "^0.6.6"
pymerkle = "^2.0.2"
pyee = "^9.0.4"
cbor2 = "^5.4.3"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
mypy = "^0.941"
pylint = "^2.12.2"
types-requests = "^2.27.14"
Sphinx = "^4.5.0"
sphinx-markdown-builder = "^0.5.5"
sphinx-autodoc-typehints = "^1.17.0"
mkdocs = "^1.3.0"
mkdocs-autorefs = "^0.4.1"
mkdocs-material = "^8.2.9"
mkdocs-material-extensions = "^1.0.3"
mkdocstrings = "^0.18.1"
pymdown-extensions = "^9.3"
eth-brownie = "^1.18.1"

[tool.poetry.extras]
novella = ["novella >=0.1.10,<0.3.0"]
docs = ["novella >=0.1.10,<0.3.0", "mkdocs", "mkdocs-material"]

[tool.poetry.scripts]
pydoc-markdown = "pydoc_markdown.main:cli"

[tool.poetry.plugins."pydoc_markdown.interfaces.Loader"]
python = "pydoc_markdown.contrib.loaders.python:PythonLoader"

[tool.poetry.plugins."pydoc_markdown.interfaces.Processor"]
crossref = "pydoc_markdown.contrib.processors.crossref:CrossrefProcessor"
filter = "pydoc_markdown.contrib.processors.filter:FilterProcessor"
google = "pydoc_markdown.contrib.processors.google:GoogleProcessor"
pydocmd = "pydoc_markdown.contrib.processors.pydocmd:PydocmdProcessor"
smart = "pydoc_markdown.contrib.processors.smart:SmartProcessor"
sphinx = "pydoc_markdown.contrib.processors.sphinx:SphinxProcessor"

[tool.poetry.plugins."pydoc_markdown.interfaces.Renderer"]
hugo = "pydoc_markdown.contrib.renderers.hugo:HugoRenderer"
markdown = "pydoc_markdown.contrib.renderers.markdown:MarkdownRenderer"
mkdocs = "pydoc_markdown.contrib.renderers.mkdocs:MkdocsRenderer"
docusaurus = "pydoc_markdown.contrib.renderers.docusaurus:DocusaurusRenderer"
jinja2 = "pydoc_markdown.contrib.renderers.jinja2:Jinja2Renderer"

[tool.poetry.plugins."pydoc_markdown.interfaces.SourceLinker"]
git = "pydoc_markdown.contrib.source_linkers.git:GitSourceLinker"
github = "pydoc_markdown.contrib.source_linkers.git:GithubSourceLinker"
gitlab = "pydoc_markdown.contrib.source_linkers.git:GitlabSourceLinker"
gitea = "pydoc_markdown.contrib.source_linkers.git:GiteaSourceLinker"
bitbucket = "pydoc_markdown.contrib.source_linkers.git:BitbucketSourceLinker"

[tool.poetry.plugins."novella.markdown.preprocessors"]
pydoc = "pydoc_markdown.novella.preprocessor:PydocTagPreprocessor"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

